@@ -2,6 +2,8 @@
 
 namespace Tests\Feature;
 
+use App\Http\Controllers\ProjectScenariosController;
+use App\Project;
 use App\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Tests\Setup\ProjectFactory;
@@ -14,12 +16,13 @@ class InvitationsTest extends TestCase
     /** @test */
     public function a_project_can_invite_a_user()
     {
-        $project = app(ProjectFactory::class)->create();
+        $this->withoutExceptionHandling();
+        $project = Project::factory()->create();
 
-        $project->invite($newUser = factory(User::class)->create());
+        $project->invite($newUser = User::factory()->create());
 
         $this->signIn($newUser);
-        $this->post(action('ProjectScenariosController@store', $project), $scenario = ['name' => 'Foo scenario']);
+        $this->post(action([ProjectScenariosController::class, 'store'], $project), $scenario = ['name' => 'Foo scenario']);
 
         $this->assertDatabaseHas('scenarios', $scenario);
     }
